More solid checking for number of iteration, and fixed some type problems with arguments.

+++ minimization.py
@@ -46,9 +46,9 @@ def check_convergence(RW, EP, Abdiff, fe_input, fe):
     return cond1 and cond2 and cond3 and cond4
 
 
-def minimize(x0, func, bounds="kurucz95",
+def minimize(x0, func, bounds="kurucz95", weights='null',
              fix_teff=False, fix_logg=False, fix_feh=False, fix_vt=False,
-             iteration=40, EPcrit=0.001, RWcrit=0.001, ABdiffcrit=0.01):
+             iteration=160, EPcrit=0.001, RWcrit=0.001, ABdiffcrit=0.01):
     """
     Sane minimization like a normal human being would do it.
     """
@@ -57,7 +57,7 @@ def minimize(x0, func, bounds="kurucz95",
     global f_teff, f_logg, f_vt, f_feh
     EPcriteria, RWcriteria, ABdiffcriteria = EPcrit, RWcrit, ABdiffcrit
     f_teff, f_logg, f_feh, f_vt = fix_teff, fix_logg, fix_feh, fix_vt
-    # Teff, logg, vt
+    # Step size in Teff, logg, vt
     step = (1000, 2.00, 1.50)
     if bounds.lower() == "kurucz95":
         bounds = [3750, 39000, 0.0, 5.0, -3, 1, 0, 9.99]
@@ -65,6 +65,9 @@ def minimize(x0, func, bounds="kurucz95",
     res, slopeEP, slopeRW, abundances = func(x0)
     Abdiff = np.diff(abundances)[0]
     parameters = list(x0)
+    if check_convergence(slopeRW, slopeEP, Abdiff, parameters[2], abundances[0]):
+        return parameters, True
+
     try:
         os.remove('minimization_profile.dat')
     except OSError:
@@ -72,9 +75,9 @@ def minimize(x0, func, bounds="kurucz95",
 
     N = 0
     while N < iteration:
-        N1 = 0
+        Nsub = 0
         cycle = [parameters[0]]
-        while (abs(slopeEP) > EPcriteria) and not fix_teff and N1 < 15:
+        while (abs(slopeEP) > EPcriteria) and not fix_teff and Nsub < 15 and N < iteration:
             # For Teff
             s = np.sign(slopeEP)
             step_i = s * step[0]/abs(np.log(abs(slopeEP)+0.0005))**3
@@ -84,23 +87,23 @@ def minimize(x0, func, bounds="kurucz95",
                 step_i = s*17625  # Maximum 17625K (half the interval?)
             parameters[0] += step_i
             parameters[0] = check_bounds(parameters[0], bounds, 1)
+            parameters[0] = int(parameters[0])
             if parameters[0] in cycle:
                 break
             else:
                 cycle.append(parameters[0])
             print_format(parameters)
-            res, slopeEP, slopeRW, abundances = func(parameters)
+            res, slopeEP, slopeRW, abundances = func(parameters, weights=weights)
+            N += 1
+            Nsub += 1
             Abdiff = np.diff(abundances)[0]
-            N1 += 1
             save_iteration(parameters)
-        N += 1
-        # TODO: Change this to use user-provided criteria
         if check_convergence(slopeRW, slopeEP, Abdiff, parameters[2], abundances[0]):
             break
 
-        N4 = 0
+        Nsub = 0
         cycle = [parameters[3]]
-        while (abs(slopeRW) > RWcriteria) and not fix_vt and N4 < 15:
+        while (abs(slopeRW) > RWcriteria) and not fix_vt and Nsub < 15 and N < iteration:
             # For micro turbulence
             s = np.sign(slopeRW)
             step_i = s * step[2]/abs(np.log(abs(slopeRW)+0.0005))**3
@@ -113,22 +116,23 @@ def minimize(x0, func, bounds="kurucz95",
             parameters[3] += step_i
             parameters[3] = 0.00 if parameters[3] < 0 else parameters[3]
             parameters[3] = check_bounds(parameters[3], bounds, 7)
+            parameters[3] = round(parameters[3], 2)
             if parameters[3] in cycle:
                 break
             else:
                 cycle.append(parameters[3])
             print_format(parameters)
-            tmp = slopeRW
-            res, slopeEP, slopeRW, abundances = func(parameters)
+            res, slopeEP, slopeRW, abundances = func(parameters, weights=weights)
+            N += 1
+            Nsub += 1
             Abdiff = np.diff(abundances)[0]
-            N4 += 1
             save_iteration(parameters)
-            if tmp == slopeRW:
-                break
+        if check_convergence(slopeRW, slopeEP, Abdiff, parameters[2], abundances[0]):
+            break
 
-        N2 = 0
+        Nsub = 0
         cycle = [parameters[1]]
-        while (abs(Abdiff) > ABdiffcriteria) and not fix_logg and N2 < 15:
+        while (abs(Abdiff) > ABdiffcriteria) and not fix_logg and Nsub < 15 and N < iteration:
             # For logg
             s = -np.sign(Abdiff)
             step_i = s * step[1]/abs(np.log(abs(Abdiff)+0.0005))**3
@@ -139,42 +143,36 @@ def minimize(x0, func, bounds="kurucz95",
             parameters[1] += step_i
             # checks bounds of logg
             parameters[1] = check_bounds(parameters[1], bounds, 3)
+            parameters[1] = round(parameters[1], 2)
             if parameters[1] in cycle:
                 break
             else:
                 cycle.append(parameters[1])
             print_format(parameters)
-            res, slopeEP, slopeRW, abundances = func(parameters)
+            res, slopeEP, slopeRW, abundances = func(parameters, weights=weights)
+            N += 1
+            Nsub += 1
             Abdiff = np.diff(abundances)[0]
-            N2 += 1
             save_iteration(parameters)
-        N += 1
         if check_convergence(slopeRW, slopeEP, Abdiff, parameters[2], abundances[0]):
             break
 
         #       Input metal...   FeI abund.
-        N3 = 0
-        while (parameters[2] != abundances[0]-7.47) and not fix_feh and N3 < 5:
+        Nsub = 0
+        while (parameters[2] != abundances[0]-7.47) and not fix_feh and Nsub < 5 and N < iteration:
             # For metalicity
             parameters[2] = abundances[0]-7.47
             parameters[2] = check_bounds(parameters[2], bounds, 5)
-            res, slopeEP, slopeRW, abundances = func(parameters)
+            parameters[2] = round(parameters[2], 2)
+            res, slopeEP, slopeRW, abundances = func(parameters, weights=weights)
+            N += 1
+            Nsub += 1
             print_format(parameters)
-            N3 += 1
             save_iteration(parameters)
-        N += 1
         if check_convergence(slopeRW, slopeEP, Abdiff, parameters[2], abundances[0]):
             break
 
-        N += 1
-        if check_convergence(slopeRW, slopeEP, Abdiff, parameters[2], abundances[0]):
-            break
-
-    print 'Stopped at %i iterations' % N
+    print 'Stopped in %i iterations' % N
     converged = check_convergence(slopeRW, slopeEP, Abdiff, parameters[2], abundances[0])
-    parameters[0] = int(parameters[0])
-    parameters[1] = round(parameters[1], 2)
-    parameters[2] = round(parameters[2], 2)
-    parameters[3] = round(parameters[3], 2)
 
     return parameters, converged
